+++ packages/fxa-geodb/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history
 
+## 1.210.1
+
+No changes.
+
 ## 1.210.0
 
 No changes.